+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Added `Tab.flash_point` from nodriver @nathanfallet
+
 ### Added
 
 ### Changed