+++ CHANGELOG.md
@@ -9,14 +9,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Added `Tab.flash_point` from nodriver @nathanfallet
-- Ensured loading is finished in expect @nathanfallet
-
 ### Added
 
 ### Changed
 
 ### Removed
+
+## [0.10.1] - 2025-07-06
+
+### Fixed
+
+- Added `Tab.flash_point` from nodriver @nathanfallet
+- Ensured loading is finished in expect @nathanfallet
 
 ## [0.10.0] - 2025-07-05